+++ Sources/MereRunCore/Quantization/QuantizedModelManifestWriter.swift
@@ -70,6 +70,7 @@ public enum QuantizedModelManifestWriter {
             case .qwen3Coder, .northMiniCode: return .code
             case .lightOnOCR: return .ocr
             case .aceStep, .magentaRT2, .muScriptor, .roFormer: return .music
+            case .apBWE: return .audio
             case .woosh, .mmaudio: return .sfx
             case .ltxVideo, .wanVideo, .cosmos3Edge: return .video
             case .psiChat: return .psi
@@ -155,6 +156,8 @@ public enum QuantizedModelManifestWriter {
                     return [.musicTranscription]
                 case .roFormer:
                     return [.musicSeparation]
+                case .apBWE:
+                    return [.audioEnhancement]
                 case .woosh:
                     return [.soundEffectGeneration]
                 case .mmaudio:
@@ -242,7 +245,7 @@ public enum QuantizedModelManifestWriter {
                 break
             case .qwen3TTS, .qwen3ASR, .parakeetASR, .sortformer, .qwen3Embedding, .openAIPrivacyFilter,
                  .qwen3Coder, .northMiniCode, .lightOnOCR, .woosh, .mmaudio, .psiChat, .deepseekV4Flash,
-                 .muScriptor, .roFormer, .inkling:
+                 .muScriptor, .roFormer, .apBWE, .inkling:
                 break
             case .aceStep, .magentaRT2, .ltxVideo:
                 manifest.defaults = MereRunModelManifest.Defaults(steps: 8, cfg: 1.0)